strbuf, merge continued lines, streams
added a lot of functionality to the strbuf and strbuf_list, also started
using file streams instead of file descriptors, for future easy
portability.

`merge_continued_lines`, now instead of mergin lines by checking chars,
we merge each line seperatly, there is still work to be done because
this is not stable, but I wanted to save it here as a checkpoint that I
can come back to in case I F up.